+++ app/page.tsx
@@ -1,7 +1,5 @@
 import HomePage from './components/pages/HomePage';
 
-export const dynamic = "force-dynamic";
-
 export default async function Home() {
   return (
     <main>
